:recycle: Refactor file and csv into file service
