+++ evaluation/src/utility/__init__.py
@@ -1,8 +1,7 @@
 from typing import List
 
 from .confirm import confirm_action
-from .csv import from_csv
-from .file import get_absolute_path, get_filename, get_files
+from .file_service import FileService
 from .logging_service import LoggingService
 from .metric import Metric
 from .model import Model, get_model_cache_dir
@@ -10,10 +9,7 @@ from .model import Model, get_model_cache_dir
 __all__: List[str] = [
   "confirm_action",
   "LoggingService",
-  "from_csv",
-  "get_absolute_path",
-  "get_files",
-  "get_filename",
+  "FileService",
   "get_model_cache_dir",
   "Model",
   "Metric"
